backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0
spacy==3.6.1
pdfplumber==0.9.0
python-docx==0.8.11
pymongo==4.4.1
redis==4.6.0
scikit-learn==1.3.0
python-dotenv==1.0.0
werkzeug==2.3.6

